third_party/WebKit/LayoutTests/http/tests/devtools/sources/autocomplete-general-expected.txt

This test checks how text editor updates autocompletion dictionary in a response to user input.


Running: testCompletionsShowUpOnKeyPress
========= Text in editor =========
name1 name2 name3 name4
name1|
======= Words in dictionary =======
[name1, name2, name3, name4]
=============

Running: testSetInitialText
========= Text in editor =========
|one two three3_
four five
a_b
six
123foo
132
seven
======= Words in dictionary =======
[a_b, five, four, one, seven, six, three3_, two]
=============

Running: testAlphaNumericWords
========= Text in editor =========
|2 2foo foo2 2foo4 foo3bar
======= Words in dictionary =======
[foo2, foo3bar]
=============

Running: testRemoveDuplicate
========= Text in editor =========
|
one
======= Words in dictionary =======
[one]
=============

Running: testSetText
========= Text in editor =========
|dog cat 'mouse';dog bird
======= Words in dictionary =======
[bird, cat, dog, mouse]
=============

Running: testSimpleEdit
========= Text in editor =========
do| cat 'mouse';dog bird
======= Words in dictionary =======
[bird, cat, do, dog, mouse]
=============

Running: testDeleteOneDogAndOneCat
========= Text in editor =========
| 'mouse';dog bird
======= Words in dictionary =======
[bird, dog, mouse]
=============
